+++ src/components/util/refundTransferClaimReturned.js
@@ -120,10 +120,8 @@ refundTransfers = async (loan, props) => {
   const createRefunds = async (account, actions) => {
     if (actions.length > 0) {
       createRefundActionTemplate(account, actions, refundClaimReturned.CREDITED_ACTION).then(
-        createRefundActionTemplate(account, actions, refundClaimReturned.REFUNDED_ACTION)
+        () => createRefundActionTemplate(account, actions, refundClaimReturned.REFUNDED_ACTION)
       );
-      const updateRecord = setPaymentStatus(account);
-      persistAccountRecord(updateRecord);
     }
   };
 
@@ -178,8 +176,13 @@ refundTransfers = async (loan, props) => {
 
     if (isAgedToLostItem || isDeclaredLostItem) {
       const accounts = await getAccounts();
-      const accountsActions = await Promise.all(
+      const updatedAccounts = await Promise.all(
         accounts
+          .map(setPaymentStatus)
+          .map(persistAccountRecord)
+      );
+      const accountsActions = await Promise.all(
+        updatedAccounts
           .map(getAccountActions)
       );
       const transferredActions = accountsActions