+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Prevent fetching resource delUser Refs UIU-2191.
 * Delete user through UI: change dialog text for no open transactions. Refs UIU-2192.
 * Delete user confirmation message. Refs UIU-2193.
+* Comment icon is missing again on Fee/Fine History page. Refs UIU-2185.
+* Add `resourceShouldRefresh` to `permissions` resource to refresh permissions. Fixes UIU-2183.
 * Add modal for `Financial transactions detail report`. Refs UIU-1960.
 
 ## [6.1.0](https://github.com/folio-org/ui-users/tree/v6.1.0) (2021-06-18)