+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 * Comment icon is missing again on Fee/Fine History page. Refs UIU-2185.
 * Add `resourceShouldRefresh` to `permissions` resource to refresh permissions. Fixes UIU-2183.
 * Add modal for `Financial transactions detail report`. Refs UIU-1960.
-* The `date picker` for `Financial transactions detail report` is cut off. Refs UIU-2204.
+* The `date picker` for report modals is cut off and user is not able select date. Refs UIU-2204.
 
 ## [6.1.0](https://github.com/folio-org/ui-users/tree/v6.1.0) (2021-06-18)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v6.0.0...v6.1.0)